facial-service/requirements.txt
gunicorn==21.2.0
Flask==2.3.2
python-dotenv==1.0.0
apscheduler==3.10.3
jwcrypto==1.5.0
dataclasses-json==0.5.14
opencv-python==4.6.0.66
onnx==1.12.0
onnxruntime-gpu==1.12.1
torch==2.0.0
Image==1.5.33
requests==2.31.0
scikit-image==0.21.0
faiss-cpu==1.7.4
coverage==7.5.0